backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
fastapi
uvicorn
timm
pillow
numpy
python-multipart
huggingface-hub